Landing Update
I changed everything to buttons.

Left with replacing the buttons with actual sprites or images.

As said earlier, the invite function doesn't work. Need to figure out the pinging and database part first.

Did not bother linking the join button to RandomRoomHandler. Could do that just for show I guess.